SERVER-95012: Reduce DataMsg batch size. (#28022)
As discussed in the ticket, changing DataMsg size from 50000 docs / 16
MB to 100 docs / 1MB.

Planning to send a separate PR to still continue to create big batches
on the sink side as that helps achieve higher throughput. Confirmed that
this change increases $merge processing time by 80%. Hoping it is fine
to temporarily incur this performance penalty.