SERVER-95873 Ensure explain command is retried when it fails on retryable errors (#28194)
Explain command is returning status `ok: 1` even in cases when the
command is killed. Due to this semantics, our current internal retry
mechanism implement in shell is not performing the retry. I have
adjusted the retry_utils.js to not consider `explain` command as
successful in case when `executionStats.executionSuccess` is `false`.