SERVER-95653 Prevent client certificate overwrite  (#28218)
This PR prevents the client certificate used for egress communication
from being incorrectly overwritten with `tlsCertificateSelector`.

Previously, when both `tlsCertificateSelector` and `tlsClusterFile` were
specified, the certificate selector was used for both ingress and egress
communication. The behavior is now corrected to check if a cluster file
was already added as a client certificate before adding the certificate
selector.