SERVER-95966 Fix SuiteSelectorConfigurationError in benchmark suite (#28227)
Co-authored-by: Daniel Moody
[daniel.moody@mongodb.com](mailto:daniel.moody@mongodb.com)

There are failures in https://jira.mongodb.org/browse/BF-35385 that
appear to be caused by the consolidated benchmark AIB component not
containing the necessary files for certain suites.

https://jira.mongodb.org/browse/SERVER-95966 mitigates the issue, but
the consolidated benchmark tests should be swapped back to using the
"consolidated-benchmarks" AIB component after the missing file issue is
resolved.