SERVER-95968 Create interface for transitively injecting a header into an idl gen rule's dependents (#28229)
Currently there isn't a way to include a header into an idl_generator
rule's dependents. This is a fairly regular process when updating idl
files that should be supported in Bazel. Update the idl_generator rule
to take in a hdrs field that will be transitively propagated to any
dependents.