SERVER-95709 Fix signal_handlers_asynchronous_test ServiceContext data race (#28003)
This change makes it so we don't destroy the global service context in
signal_handlers_asynchronous_test. The signal processing thread makes
unsynchronized reads from the global service context, so destroying it
can cause a race condition. Most tests do not have this thread active.

As a result of this test starting the signal processing thread globally
(with no way to stop it), I've also moved the signal handler tests into
their own binary. This prevents the effects of that thread from leaking
into other tests.

---------

Co-authored-by: Ryan Berryhill <ryan.berryhill@mongodb.com>