SERVER-95733 Handle rare type of error correctly in testing (#28074)
There is a rare LDAP failure case whose cause is unknown, where
occasionally, a call to rebind that should have failed with an error
will return no results but give a successful status. Every time we call
LDAP search, a return of no results will immediately cause an error, so
this won't result in any unexpected behavior. However, this kind of
failure will not be retried, while a normal rebind failrue wil. For that
reason, we turn off retries in this test so that the normal failure and
the uncommon failure will look the same.