SERVER-94939 Add tests for deep stack traces (#28315)
This change adds new tests checking that deep stack traces work (on
non-windows platforms). I slightly refactored the `callNext` function
from `stacktrace_libunwind_test` and pulled it out into a new TU we can
use in both that test and `stacktrace_test`. I added two tests that
check the behavior for stacktraces close to the 100-frame limit: one
where we are under the limit and check that all expected frames are
present, and one where we're over the limit and check that at least a
certain number of the expected frames is present.

These tests are not enabled on Windows because we use inline assembly to
prevent tail call optimization from happening in order to generate large
stack frames, and MSVC doesn't allow inline assembly on some platforms,
including x86-64.