SERVER-96017 Add version_impl dependency to signal_handlers_test (#28308)
This test was recently split off from the very large `util_test`. This
is because it starts the signal processing thread, which can't ever be
stopped, and that's a detail we don't want to leak into unrelated tests.
However, when it was split, the dependency on `version_impl` didn't
follow it. This test needs the `EnableVersionInfo` initializer from that
library since the signal processing thread accesses the version info
under certain conditions when receiving SIGUSR1. As a result, the
missing dependency makes the test hit an invariant failure some of the
time.