SERVER-96047 move linters out of scons (#28323)
It turns out our linters in scons were doing compiles and take like ~20
mins right now in the commit queue and on the waterfall. In extreme
cases they would time out. I am not sure how long this has been an issue
but this brings them back down to 1m.