SERVER-83475. Extend notablescan coverage to sharded clusters and replicasets (#28180)
`notablescan` is a server parameter we offer users to configure and it
is not covered in different topologies. Adding it to the mutational
fuzzers so it gets all its coverage at no extra cost.

Correctness of this parameter is already well covered in its own units
and jstests (`jstests/noPassthrough/query/*notablescan.js`)