SERVER-96123: Remove dead Bonsai explain code (#28401)
This PR is part of the Bonsai removal effort. It deletes some ABT
explain functionality that was specific to Bonsai, including the
"user-facing explain" generation (intended to be used in queryPlanner
explain output), other explain versions that are unused by QE tests
today, and explain generation for properties. It also removes some
explain-related test helpers.