SERVER-96030 Make interface parity test join metadata refresh (#28343)
Some test cases block replication to force a write concern error, and
check there is a `writeConcernError` field. This field is added at the
service entry point's final wait for write concern just before returning
the response. However, the command might fail before reaching that point
if an asynchronous filtering metadata refresh, triggered by collection
creation during test setup, did not finish by the time replication is
stalled. This is because the refresh will be stalled waiting for
replication, and the user command will wait for the refresh, and fail
because replication is stalled.

To make the test deterministic, this patch adds usage of getShardVersion
to determine if the refresh is finished before stalling replication.