SERVER-93167: Remove obsolete Bonsai tests, refactor other tests to use Expressions instead of Nodes (#28443)
This PR is part of the Bonsai deletion effort. It audits the unit tests
in optimizer_test and reference_tracker_test to see which ones make
sense to delete, which make sense to keep as-is, and which make sense to
keep with modifications (for example, with modifications to operate over
Expressions instead of Nodes and Paths).