SERVER-96136: Move ScopedDebugInfo logging after stack trace logging (#28484)
Before this change, when an invariant or tassert is tripped, we logged
the assertion failure, the ScopedDebugInfo, and the stack trace -- in
that order. ScopedDebugInfo does its stringification lazily, and it
executes code that could error/crash at logging time. With this log
ordering, there is a risk that something could go wrong during
ScopedDebugInfo logging, prohibiting us from logging the stack trace.
This change flips the order so that stack trace logging, which we expect
to be the safer/less-likely-to-go-wrong step, happens before
ScopedDebugInfo logging.