SERVER-96193 Set timeout code in AsyncOpTimeoutWithOpCtxDeadlineLater to ensure consistent failure (#28493)
AsyncOpTimeoutWithOpCtxDeadlineLater sets a request timeout but does not
set a request timeout code. As a result, the timeout code returned from
startCommand will context-dependent. In particular, if a timeout happens
to be hit during connection acquisition (which is more likely on slower
hosts like the sanitizer ones), a
PooledConnectionAcquisitionExceededTimeLimit will be returned rather
than the expected NetworkInterfaceExceededTimeLimit error.

To solve this, commit sets a timeout code on the request so that it will
be returned during any timeout, regardless of what stage the command
execution is at when the timeout occurs.