SERVER-96163 Use typed fields where possible in FLESchema structs (#28455)
Any type in `fle_field_schema.idl` currently using `bindata_generic`
(with the exception of FLE2InsertUpdatePayloadV2.value` which will be
handled later) is moved to an explicit IDL named type from
`fle_tokens.idl`.

Most of this change is as simple as updating the `type:` field in the
IDL file, then changing get/set calls to work with the concrete type
rather than generic ConstDataRange buffers.

The big change in moving `EncryptedStateCollectionTokensV2` into
`StateCollectionTokensV2` as a POD type with a single active method
`encrypt(ECOCToken)` which will produce
`StateCollectionTokensV2::Encrypted` with the type we can
serialize/deserialize for the actual payload types. The Encrypted
subclass also has its own `decrypt` method to devolve back to the
unencrypted `StateCollectionTokensV2` class. This allows us to treat the
two states of this data as discrete types with discrete purposes.