SERVER-92824 Prevent incorrect $addFields $match swap (#27411)
$addFields 'a.b.c' modifies not only 'a.b.c' but also 'a.b' and 'a': it
turns any scalars or missing fields into objects to make room for
'a.b.c'. This means a subsequent $match on 'a.b.x' may not be
independent: the $addFields causes the $match path to visit some missing
fields it otherwise wouldn't.

This means we have to inspect the $match predicate more carefully to
decide whether the swap is correct. Details and examples are documented
in 'jstests/aggregation/sources/addFields/independence.js'.