SERVER-94991 Make ValidateCollections faster (#28537)
The last attempt failed due to resmoke_end2end_tests failing on debug
variants. This is because the error code we used in our failpoint
expected extra info and on debug variants it crashes the server with a
fatal assertion when this extra info does not exist. I switched the
error code to one that does not have this extra info requirement.