SERVER-96279 Make resmoke easier to run on spawnhosts (#28592)
On spawnhosts there is not enough git history and resmoke fails when
trying to run `git describe`. In CI, resmoke normally does not need git
to be active and uses a file it downloads from the artifacts tarball to
get this info. I made spawnhosts also use this tarball to get this info
so now resmoke runs fine right after the spawnhost script completes.