SERVER-96280 Make local-resmoke-invocation.txt errors non-fatal in alternate evergreen projects (#28594)
The local-resmoke-invocation.txt file generation has some logic that is
heavily on the evergreen yml configuration we have in the
mongodb-mongo-* projects. Other projects such as dsi and mongosync who
use resmoke might not have these same configuration implementations that
resmoke relies on for this file. In those projects these errors should
appear as warnings and not make resmoke end abruptly.