SERVER-93528 Remove flaky test for crash without --repair (#28492)
This test failed in [BF-34529](https://jira.mongodb.org/browse/BF-34529)
due to slow process termination on a windows system, leading to the
expected abortion not being detected in time. It looks like the other
test `runRepairAndVerifyIndexIsRepaired` is the one actually exercising
and verifying the intended behavior we want to test for, since we are
looking for candidates to trim for test flakiness, I'm opening this to
discuss whether this test brings value.