SERVER-96203 Wait until task has been canceled before exiting scheduleIntoPoolSpinsUntilThreadPoolTaskExecutorShutsDown block (#28502)
This fixes a race in the ShutdownAndScheduleWorkRaceDoesNotCrash test.
Since cancellation during shutdown happens outside of the lock, it was
possible for the task about to be scheduled to successfully run before
the shutdown thread had a chance to cancel it. This guarantees the task
will be canceled by the time it is scheduled onto the pool.