SERVER-94834 Centralize FLE processing for commands (#28298)
Centralize FLE processing for commands.

There was a lot of code repetition for FLE processing in all affected
commands.
For example, every FLE-aware command checked
* if the incoming request contained encryption information
* disabled the query diagnostics if so, so that no sensitive field
information could be leaked into the diagnostics.
* checked if the requested was already processed for FLE
* and if not, did the actual, command-specific FLE processing

The repeated code is now moved into a central function
`prepareForFLERewrite`, which returns `true` if the caller is supposed
to perform the command-specific FLE-processing. It will return `false`
in case the request does not need to be FLE-processed.
The function will also disable query diagnostics if necessary.