SERVER-96166 Fall back to repair=true on importCollection error (#28509)
importCollection can fail if during the dry run phase more than two
checkpoints run, invalidating the storageMetadata that is passed in.
When running the production import, passing in `panic_corrupt=false`
will allow us to catch the error and retry with `repair=true`. This uses
the latest state of the table, but we only want to use it as a fallback
as it reads the entire table.