SERVER-90843: Remove explicit StringData casts in args to setters (#28485)
With the recently added capability to set IDL string fields via a method
that uses std::move, it's best practice to not pass StringData objects
(which will inhibit the move due to StringData being a non-owning
reference). This is especially true in cases where the string value
passed to the setter is a specific instance of a rvalue, such as an
overloaded operator expression. Thus, I've removed all the identified
instances (as mentioned in the ticket) of explicit StringData
conversions.

Ex: not as crucial
`indexSpec.setName(StringData(clusterKey)); // where clusterKey is of
type StringData`

Ex: inhibits a move
`indexSpec.setName(StringData(clusterKey + "_1")); // where clusterKey
is of type StringData`