SERVER-88000 Use better backoff for writeConflictRetry loops (#28070)
The existing retry strategy of a truncated linear backoff (which the
comment incorrectly claimed was truncated exponential backoff) was
surprisingly decent, and for a lot of operations doing any backoff at
all only hurts performance. However, it does have some cases where
performance fell off a cliff at high thread counts, which this backoff
strategy aims to mitigate while not hurting the more common scenarios.

The minimal microbenchmark shows significant benefits from rapidly
scaling up the backoff, but this does not appear to generalize to any
more realistic workloads. This is testing a case where close to 100% of
the runtime is spent in committing the write, and introducing a query
before the write changes the performance characteristics. A find and
update command processes most of the write conflicts during the plan
execution stage rather than committing, so even though each failed
commit becomes more expensive with a longer query, there ends up being
significant backoff along the way.

This introduces a backoff strategy of (10% of elapsed time per attempt)
* 1.1^attempt. Scaling the base factor to how expensive the operation
being retried is necessary because some of the operations being retried
can be performed 150k times per second while others can be done 100
times per second, and any static backoff amount useful for the latter
destroys the performance of the former. The extremely gentle curve is
experimentally optimal for not regressing the performance of any tests
while still improving some (but not improving those which do benefit
from exponential backoff as much as possible).

---------

Co-authored-by: Billy Donahue <billy.donahue@mongodb.com>