SERVER-96468 Remove ScopedDBConnection use from cluster killOp (#28728)
After removing this and a few other usages, we can delete the 'legacy'
connpool from the codebase. We can simply re-use the `shard` object we
have in scope rather than creating a new connection.

Note that we should use PrimaryOnly read-preference as that is the
default read-preference that the DBClientRS infrastructure selects in
the absence of an explicit read-preference; it will attempt to parse
using ReadPreferenceSetting::fromContainingBSON which uses PrimaryOnly
as the default read-preference when none is provided. This behavior
matches the code we're removing but is not necessarily correct; fixing
that bug is tracked in SERVER-96446.