SERVER-75546 Use rawMongoProgramOutput instead of checkLog in election_handoff.js (#28690)
During election handoff tests, we rely on checkLog to make sure the
correct node has stepped up. But it's possible for important logs to be
truncated if we hit the 16MB size limit of a BSON object. We switched to
use rawMongoProgramOutput, which is more robust, so we can filter only
for the logs we care about.