SERVER-95257 expose a command to untrack from the sharding catalog unsplittable collections residing on the db primary shard (#28097)
Exposing to the user a command to untrack a collection from the sharding
catalog on the config server when the collection resides on the primary
shard.
Extending several tests to guarantee maximum coverage for the command. 
Note the command won't be documented for the user. The command is meant
for TSE only.
Usage:
`db.adminCommand(untrackUnshardedCollection: nss)`