SERVER-96477 Remove ScopedDBConnection from cluster aggregation planner (#28742)
Replaces the use of a ScopedDbConnection to run a listCollections
command with the RPC functionality built into Shard.

After removing this and a few other usages, we can delete the 'legacy'
connpool from the codebase. We can simply re-use the shard object we
have in scope rather than creating a new connection.