SERVER-95048 Rename HostTypeRequirement::kMongoS into kRouter (#28792)
This PR continues the work in
[SERVER-85808](https://jira.mongodb.org/browse/SERVER-85808) and
[SERVER-94989](https://jira.mongodb.org/browse/SERVER-94989). kMongoS
value is being renamed to kRouter in HostTypeRequirement enum. Since
HostTypeRequirement enum values are closely related to mergeType field
in $_internalSplitPipeline and explain(), mongos is being renamed to
router on these fields.

Therefore:

* HostTypeRequirement::kMongoS is being renamed to kRouter
* mergeType value on $_internalSplitPipeline is being renamed from
'mongos' to 'router'
* mergeType value on explain results is being renamed from 'mongos' to
'router'

Note: #28406 was reverted because of some newly added tests that I
missed to update.