SERVER-96630: Fix copying and concurrency issues in sharded backup test. (#28788)
# Issue

[SERVER-96630](https://jira.mongodb.org/browse/SERVER-96630)

# Description

We have observed failures in the Sharded backup tests, where the oplog
does not seem to contain a complete list of all the transactions.
Inspecting those backups shows wiredtiger journal files with zero
length. Unfortunately, we're not sure why.

To assist in debugging this issue should it occur again we make a few
changes
- ensure that copies of the initial backup complete before copies of the
extended backup begin (i.e. to avoid races)
- ensure that the backup cursor remains open until all copes have
completed (WT says to do this in their API)
 - make copy failures an error (previously they were silently ignored)