SERVER-95377: Make test robust to network timeouts (#28585)
`catchup_ignores_old_heartbeats.js` previously may result in rollback on
the old primary if the heartbeat to the new primary times out, and the
new primary does not update its target opTime as part of catchup. This
change waits for the heartbeat to be received by the new primary before
starting replication, which should ensure that the new primary fully
completes catchup up until the target optime.