SERVER-95597 Change mirror_reads.js to account for slow executor (#28841)
For mirror_reads.js, if there are reads that are scheduled via
ExecutorFuture but not run, it's possible for
`sent === resolved` and `seen !== sent` because the `resolved` and
`sent` metrics are incremented using the executor. If this executor
doesn't run in time, then certain assertions in mirror_reads.js will
fail.