SERVER-96064: Optimize away metadata $sort directly after $vectorSearch for single node environments  (#28699)
[SERVER-96064](https://jira.mongodb.org/browse/SERVER-96064)

This PR introduces an optimization that removes a `$sort` stage the
comes directly after a `$vectorSearch` stage, where the `$sort` and
`$vectorSearch` sort criteria match as the `vectorSearchScore`. In this
case, the input and output of the `$sort` stage would be the same and
thus redundant.

This optimization is important for the desugar of $rankFusion, as this
case will apply and we would like to remove the unnecessary `$sort` to
improve performance. In the desugar case, the redundant `$sort` will
come from the desugar of a `$setWindowFields` stage - however the
optimization logic is agnostic.

This PR only handles single node environments, because in sharded
environments, the `$sort` will come after a `$mergeCursors` stage on
mongos, and thus is essentially a different implementation and will be
handled separately.

Note: unit tests (just jstests) are added to this PR because unit tests
with $search/$vectorSearch pipelines are not yet set up (see
[SPM-3850](https://jira.mongodb.org/browse/SPM-3850) which is planned to
enable this)