SERVER-95867 Cardinality estimation driver with heuristic CE (#28612)
Implement a simple driver to compute CE of arbitrary QSN plans. The
driver walks recursively a QSN tree and invokes CE for each node
bottom-up.

- The CE computation of conjunctive nodes uses a stack of conjuncts'
selectivities. This allows to encode selectivities of implicit and
explicit conjunctions in the same way.
- The driver calls heuristic CE and is integrated with the current QSN
cost estimation.

The driver is simple in the sense that it doesn't use CE caching and
logical node equivalence.