SERVER-96692 - Export collection from RS to avoid alterMetadata (#28914)
This test can occasionally fail if enough unlucky EBUSYs are returned by
WiredTiger. This is more common after SERVER-95921 as checkpoint() is
called after dry-run on each collection.

We can't avoid this with synchronization on the MongoDB side as WT can
still do checkpoints in the background, so there's no way to 100%
guarantee success.
Here we can avoid the issue by ensuring the table logging settings don't
need to be changed. Currently these are the calls that are failing due
to conflict with checkpoints, so if we never call them, we don't have
problems. This was the strategy used in SERVER-95921, but this test
missed out since it doesn't rely on the same helper function