SERVER-95855 applyOps of an applyOps of a dropDatabase causes invariant failure that crashes mongod (#28736)
Fixes a MongoDB server crash when running a `dropDatabase` command
embedded inside two (or more) `applyOps` commands, for example:

```js
db.adminCommand({
	applyOps:[
		{op:"c",ns:"admin.$cmd",o:{
			applyOps:[
				{op:"c",ns:"test.$cmd",o:{
					"dropDatabase":1
				}}
			]
		}}
	]
})
```

Generally, running commands on `applyOps` requires taking the global
write lock. However, `dropDatabase` is a special case as it may await
replication internally.
[SERVER-62759](https://jira.mongodb.org/browse/SERVER-62759) allows a
single `dropDatabase` command to run on `applyOps`, by checking for this
case and releasing the global write lock acquisition.

However, it is possible to nest an `applyOps` command inside another
`applyOps` command which in turn runs `dropDatabase`. In this case, both
the outer and inner `applyOps` will acquire the global write lock (which
by itself is fine as it is a recursive lock). The check from
[SERVER-62759](https://jira.mongodb.org/browse/SERVER-62759) will
release the inner global write lock acquisition, but it has no knowledge
of the outer global write lock acquisition, so the global write lock is
still taken when running `dropDatabase`, which triggers an invariant.

As there isn't a use case for this nested `applyOps` use, we fail the
operation and return an error the the user. We detect it by checking if
the global write lock is still taken after releasing the acquisition. If
it is, an outer `applyOps` is also holding the lock and we can not
proceed.