SERVER-82720 Refactor run_aggregate lock acquisition and collation resolution (#28752)
These changes create a new class, `AggCatState` to encapsulate
catalog-related state in `runAggregate()`.

Since a lot of the complexity in the code was related to different types
of pipelines, I made three subclasses of `AggCatState` to abstract away
the differences:
- `CollectionlessAggCatState`
- `OpLogAggCatState`
- `NormalAggCatState`

These changes also covers the work of SERVER-82719, which I hadn't
realized was actually separate issue from SERVER-82720. Will's POC
branch provided lots of helpful context here!

I opted for a separate class instead of using `AggExState` here since I
wanted to adhere to the spirit of @joe-mongodb's change in #25450 and
avoid variables that are in use in some code paths and unset and unused
in others. For example, collectionless pipelines don't need to take out
catalog locks at all, and (I hope) having a dedicated subclass for that
case makes that clearer to the reader.

This new class encapsulates the state of catalog-related concerns
(including locks, which will be freed by the destructor), as well as the
collection collation, which must be resolved with any specific collation
provided as part of the request. This makes several of the function
calls in `runAggregate()` more concise since they have much fewer
parameters with this change.