SERVER-96256: switched to unordered_map and added a new metric (#28569)
Heartbeat handles are stored in the
[queue](https://github.com/10gen/mongo/blob/cbfbc527dd8da88513dc596a6913f6b1eefc0e96/src/mongo/db/repl/replication_coordinator_impl.h#L1917)
while heartbeats are being executed and discarded after heartbeats are
handled. This PR changed the way the handles are stored so removal will
take a constant time and linear(that is important because linear search
is performed while the main lock is acquired. An auxiliary metric was
added to show the max number of enqueued handlers seen by any given
moment as an additional indicator of the system health.
https://jira.mongodb.org/browse/SERVER-96256