SERVER-87065: Re-enable time-series validation checks in testing (#29014)
This is a second attempt to re-enable time-series validation checks that
had been disabled in testing when we made the change to switch
time-series validation failures from warnings to errors. The two checks
re-enabled here are:

That the timestamp embedded in the OID matches that of the control.min
field
That the control.min.timestamp field is >= the observed minimum
timestamp, rounded down for granularity
For some tests, this required updated the manually created buckets that
had been created in testing to pass validation - for example, some
manually created buckets had embedded timestamps that did not match
their control.min field/the observed minimum timestamps.

The differences between the previous attempt and this on are that:
1) For the embedded timestamp check, I used the
timeseries::collectionMayRequireExtendedRangeSupport function directly.
In the first attempt I checked the value of the
getRequiresTimeseriesExtendedRangeSupport flag, but in a patch I had a
test where this flag value was set to false despire there being an
extended range measurement in the collection. This is a bit suspicious,
and I will file a ticket to look more into this because ideally I think
we should use that flag, but to get a working check in I have for the
time being used the collectionMayRequireExtendedRangeSupport directly.

2) When comparing the control minimum data for non-timestamp data with
the observed minimum , I made the comparison ignore field order. There
was a fuzzer failure where the measurements differed only in their field
order.

3) When comparing the control.max and observed max timestamps for
buckets that have pre-1970 events, relaxed the strict equality check,
since the control.max can end up rounded up to the epoch and be greater
than the observed maximum value.