SERVER-92427 Fix QueryPlannerIXSelect::rateIndices for nodes that can… (#26908)
…not use index under $not

Currently index tagging is inconsistent for several types of nodes under
$not operator, e.g. {a: {$not: {$mod: [2, 0]}}}.
If the expression is unnormalized, it is expanded by parsing into {a:
{$not: {$and: {$mod: [2, 0]}}}} and recursive calls to rateIndices
function lead to tagging of $mod and eventual index on the path 'a'.
This is incorrect since the index cannot be used under $not operator.
In case of normalized expression, the $and is removed and the code path
goes through a special handling of $not and its direct child. In this
case the nodes are tagged, but the eventual index is not considered
compatible.

This patch passes down information about the presence of $not to the
recursive calls of rateIndices() to avoid incorrect tagging of indices.