SERVER-89138 Extend namespace limit for unsplittable collections to 255 characters (#28181)
An unsharded collection has a nss limit of 256. However, once tracked,
the collection will be limited to 235. This is a strong limitation, as
every collection beyond that limit can't be moved as consequence.
The PR proposes the following:
- Extends the limit of 235 to 255 for unsplittable collection (unsharded
collections registered in the cluster catalog)
- Extend the limit on any internal collection to 512. This allows
config.cache.chunks.<collection_name> to pass the current soft
limitation and possibly to future collections.