SERVER-95859 Fix -Wmissing-prototypes warnings and enable it (#28170)
-Wmissing-prototypes warns when a non-static, non-inline definition of a
function is found without a matching declaration. This typically means
that a helper function has been unintentionally exported rather than
being made static or put in an anonymous namespace, which results in
slightly worse codegen (static functions called from exactly one place
will be inlined much more aggressively) and hurts link times. The other
causes I ran into while fixing this were:

1. Missing header includes. A few source files just didn't include their
corresponding header.
2. Signal handlers, which because of z/OS giving C++ functions a
different calling convention have to be `extern "C"`.
3. Functions shared between multiple source files but intentionally not
declared in any headers.
4. Functions invoked via dysym() or equivalent
5. Test helpers deliberately doing weird things (such as in the tests
for backtraces)

This was mostly a pretty mechanical process: for each warning, check if
a header was missing and if so include it. Check if something clever is
going on, and if so add a declaration immediately before the definition
to silence the warning with no functional changes. If ADL is being used
(e.g. operator overloads and some visitor functions), make the function
`inline`. Otherwise wrap it in an anonymous namespace (and increase the
bounds of that namespace as much as possible).

Doing this revealed that many of the unintentionally-exported helper
functions were actually just dead code, so I made a pass over the
-Wunused warnings and deleted a few thousand lines. I did not attempt to
make the code base fully -Wunused clean.

Confusingly GCC calls this flag `-Wmissing-declarations`, while with
clang that's an unrelated warning. Since that name is ambiguous I've
used the `-Wmissing-prototypes` name