SERVER-91974 Remove balancer force reload of shard registry (#29021)
The Balancer is a replica set aware service, and only runs in the config
server primary (step downs cause termination). The config server is the
source of the TopologyTime, and all ShardRegistry calls within the
balancer use the causally consistent methods. Thus, forcing a reload in
each balancing round is unnecessary.