SERVER-85172 Fix test assumptions about getMore data availability (#28931)
Tests must not assume that a call to getMore on the server will always
produce a batch with data in it.
It is possible that the call on the server does not get a lot of CPU
time before the 1s timeout for the getMore command elapses.
In that case the server may return an empty batch for a getMore call,
and client code must be able to deal with this.

That means clients cannot assume that a call to `cursor.next()` produces
data.
Such calls should be preceded in tests with `assert.soon(() =>
cursor.hasNext());` to make sure that data is actually available.

---------

Co-authored-by: Denis Grebennicov <denis.grebennicov@mongodb.com>