SERVER-82721 Refactor run_aggregate to use a helper function that prepares the pipeline (#29069)
[Ticket](https://jira.mongodb.org/browse/SERVER-82721)

This ticket comes after all the nice state encapsulation that Joe and
Chris did in run_aggregate, which made this work a lot cleaner! This
ticket builds on that work by cleaning up the high level path with a
helper function that prepares the pipeline for execution. It's super
close in spirit to Will's very helpful POC.

After this,
[SERVER-82722](https://jira.mongodb.org/browse/SERVER-82722), and
[SERVER-82723](https://jira.mongodb.org/browse/SERVER-82723), the main
logical steps of run_aggregate should be clearer: 1) preparation of the
parsed pipeline 2) preparation of the executors, 3) actually performing
the execution.

---------

Co-authored-by: Christopher M. Wolff <chris.wolff@mongodb.com>