SERVER-96996 Add summary of failing queries to git diff output (#29195)
This PR adds some metadata about failures captured by `git diff`. It
contains logic to extract the failing test id from a hunk header, which
is then mapped to a corresponding query. After the git diff output of
one failed test, we will output:

![Screenshot 2024-10-30 at 3 00
16 PM](https://github.com/user-attachments/assets/2011d60f-e398-417a-b594-f99bcbb655c7)

After all tests are run, there is an additional Failure Summary
describing:
- Number of failed files
- Number of failed queries
- A list of filepaths associated with the failed files

![Screenshot 2024-10-30 at 3 00
43 PM](https://github.com/user-attachments/assets/0111d572-c9f3-4389-abb7-179a75b69619)

Note that this ticket was ported over from
[DEVPROD-12412](https://github.com/10gen/QueryTester/pull/43). It also
includes a fix to conditionally include headers depending on whether
we're in a unix or windows-based system
([BF](https://jira.mongodb.org/browse/BF-35683)), as well as a fix to
formatting.