SERVER-93874 Only use time-series control.max bucket metadata when the date is after unix epoch (#26614)
The root cause is query code not accounting for SERVER-67816, which can
cause wrong results when we try to use control.max when the bucket has
data <1970. This change makes two fixes:
- Don't use bucket-only plans (plans that only access the bucket
metadata) if we need the control.max field for time.
- Block processing will check the date before using the bucket
control.max for time.

I was going to include the PBT change that caught this, but it also
catches two more TS bugs and will have a lot of TODOs so it seems
cleaner to post it separately.