SERVER-82674 Prevent NITL from owning a TransportLayerManager (#28694)
This commit prevents the NetworkInterfaceTL from creating a
TransportLayerManager that it owns. In order to startup the
NetworkInterfaceTL, it is now a requirement to either pass in a service
context or have a service context set globally. In both cases a
TransportLayerManager must also be set on the service context.