SERVER-96215: Max queue depth parameter added to ticket holder (#29200)
To be able to report to the user of mongod that a server is overloaded,
we need to limit the amount of waiting admissions to the server. A new
parameter has been added to the ticket holder to set this maximum, and
it will throw if the max is surpassed.