SERVER-94944 Add mechanism to test EINTR resilience (#28429)
This change adds a way to start a thread intended to test our resilience
to system calls being interrupted by signals. The thread sleeps for a
randomized interval and then picks another thread in the process at
random to send a signal to. It sends `SIGRTMIN` to the thread, which is
mapped to a no-op handler. This has the effect of interrupting any
interruptible running system call.

A few details worth noting are:
1. If the signal's disposition is `SIG_IGN` (rather than an explicit
no-op handler), the signal call seems to get restarted transparently.
2. Using SIGSTOP/SIGCONT without explicit handlers behaves similarly.

In both cases above, some system calls might `EINTR` out, but not all
that can return `EINTR`. Using an explicit no-op handler `EINTR`s out
more system calls.

We use `SIGRTMIN` because the meaning of realtime signals is entirely
application-defined. The properties of realtime signals in contrast to
non-realtime signals isn't important to us here.

This change just adds the mechanism but doesn't actually use it yet.

---------

Co-authored-by: Alex Li <alex.li@mongodb.com>